requirements.txt
boto3
python-dotenv
strands-agents
streamlit
langchain
langchain-community
faiss-cpu
langchain-text-splitters
langchain_aws
